Uplevel DEPS to reflect dependent merges
Conflicts:
	.DEPS.git
	.gitmodules
	base/compiler_specific.h
	build/android/envsetup_functions.sh
	build/android/gyp/write_ordered_libraries.py
	build/common.gypi
	chrome/chrome_android.gypi
	chrome/chrome_tests_unit.gypi
	content/content_tests.gypi
	sandbox/linux/sandbox_linux.gypi
	skia/skia_library_opts.gyp

Former-commit-id: ae41f478025ebc99755d07d36d5e773f6546b84f

+++ chrome/browser/android/omnibox/omnibox_prerender.h
@@ -5,7 +5,7 @@
 #ifndef CHROME_BROWSER_ANDROID_OMNIBOX_OMNIBOX_PRERENDER_H_
 #define CHROME_BROWSER_ANDROID_OMNIBOX_OMNIBOX_PRERENDER_H_
 
-#include "base/android/jni_helper.h"
+#include "base/android/jni_weak_ref.h"
 #include "base/memory/scoped_ptr.h"
 #include "base/strings/string16.h"
 
@@ -50,7 +50,7 @@ class OmniboxPrerender {
                       jobject obj,
                       jstring j_url,
                       jstring j_current_url,
-                      jint jsource_match,
+                      jlong jsource_match,
                       jobject j_profile_android,
                       jobject j_tab);
